Fixed more unnecessary typings

+++ Gengar/Handlers/InteractionHandler.cs
@@ -98,7 +98,7 @@ namespace Gengar.Handlers
 
             foreach (var user in birthday.ToList())
             {
-                var userInGuild = Guild.GetUser((ulong)user.Userid);
+                var userInGuild = Guild.GetUser(user.Userid);
 
                 if (userInGuild == null)
                 {
